--- LinkedBagDS/EventTicketMain.cpp
@@ -1,116 +1,0 @@
-#include <iostream> 
-#include <string>
-
-// TO DO: #include any other libraries you need
-#include "EventTicket340.h"
-#include "Organizer.h"
-#include "Event.h"
-
-
-using namespace std;
-
-/** 
- * 
- * Displays the application's main menu
- * pre create a new object of type Organizer
- * @param organizer object to interact with
- * 
- * */
-void displayOrganizerMenu(Organizer& organizer){
-	int organizerChoice = 0;
-	do {
-		cout << "\n Hi, "<< organizer.getUsername() <<", what would you like to do:\n"
-		<< "1. Display Profile\n"
-		<< "2. Modify Password\n"
-		<< "3. Create Event\n"
-		<< "4. Display All Events\n"
-		<< "5. Display Kth Event\n"
-		<< "6. Modify Event\n"
-		<< "7. Sell Tickets\n"
-		<< "8. Delete Event\n"
-		<< "0. Logout\n"
-		<< "Choice: ";
-		cin >> organizerChoice;
-
-		switch (organizerChoice) {
-			case 1:{
-				// TO DO: display organizer's profile information
-				//      : e.g. organizer.displayProfile();
-				break;
-			}
-			case 2: {
-				// TO DO: ask for new password and update organizer's password
-				break;
-			}
-			case 3: {
-				// TO DO: ask organizer to choose event type, then ask them to input event details.
-				// Create the event and add it to the organizer's events
-				break;
-			}
-			case 4:{
-				// TO DO: display all organizer's events
-				//        You may re-use code from class demo
-				break;
-			}
-			case 5: {
-				// TO DO: ask the organizer for a value k
-				// Find the Kth event, if k > LinkedBag size, 
-				//    return an error message that includes the size of the LinkedBag
-				break;
-			}
-			case 6: {
-				// TO DO: ask the organizer for the index of the event they want to modify.
-				// Find the event, then prompt them for the new name and description.
-				// Modify the event accordingly. 
-				// If index > LinkedBag size, 
-				//    return an error message that includes the size of the LinkedBag
-				break;
-			}
-			case 7: {
-				// TO DO: ask the organizer for the index of the event 
-				// Find the event, then sell the tickets. 
-				// If index > LinkedBag size, 
-				//    return an error message that includes the size of the LinkedBag
-				break;
-			}
-			case 8:{
-				// TO DO: ask the organizer for the index of the event they want to delete 
-				// Find the event, then remove it from the list. 
-				// If index > LinkedBag size, 
-				//    return an error message that includes the size of the LinkedBag
-				break;
-			}
-			case 0: {
-				cout << "Logging you out." << endl;
-				break;
-			}
-			default:
-				cout << "Invalid choice. Please try again." << endl;
-		}
-
-	} while (organizerChoice != 0);
-}
-
-
-int main(){
-	// Instantiating the program using the default constructor
-	// With this implementation, the application will only have one organizer
-	EventTicket340 eventTicket340; 
-
-	cout << "\n Welcome to EventTicket340:" << endl;
-	// TO DO: Ask the organizer to enter their information 
-	//        Instantiate a new Organizer object
-
-
-	// call eventTicket340 createOrganizer function 
-	// replace /*...*/ with the right parameters
-	eventTicket340.createOrganizer(/*...*/);
-
-	// Retrieve the organizer 
-	Organizer currentOrganizer = eventTicket340.getOrganizer();
-
-	// Display the main menu
-	displayOrganizerMenu(currentOrganizer);
-				
-	return 0;
-}
